refactor: tokenProperty 위치를 config 패키지로 이동
