+++ backend/src/main/java/com/woowacourse/gongseek/auth/config/RefreshTokenProperty.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.auth.domain;
+package com.woowacourse.gongseek.auth.config;
 
 import io.jsonwebtoken.security.Keys;
 import java.nio.charset.StandardCharsets;
