+++ backend/src/main/java/com/woowacourse/gongseek/auth/config/TokenProperty.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.auth.domain;
+package com.woowacourse.gongseek.auth.config;
 
 import java.security.Key;
