+++ backend/src/main/java/com/woowacourse/gongseek/auth/presentation/AuthController.java
@@ -52,5 +52,4 @@ public class AuthController {
 
         return ResponseEntity.ok(new AccessTokenResponse(tokenResponse.getAccessToken()));
     }
-
 }
